starting slow migration to Django 1.10
