+++ urls.py
@@ -1,38 +1,37 @@
-from django.conf.urls import patterns, include
+import views
 from django.conf import settings
-
-# Uncomment the next two lines to enable the admin:
+from django.conf.urls import include, url
+from django.conf.urls.static import static
 from django.contrib import admin
 admin.autodiscover()
 
-urlpatterns = patterns('',
-    (r'^media/(?P<path>.*)$', 'django.views.static.serve', {'document_root': settings.MEDIA_ROOT }),
-    (r'^$', 'equilibrator.gibbs.main_page.MainPage'),
-    (r'^about', 'equilibrator.gibbs.info_pages.AboutPage'),
-    (r'^faq', 'equilibrator.gibbs.info_pages.FAQPage'),
-    (r'^whats_new', 'equilibrator.gibbs.info_pages.WhatsNewPage'),
-    (r'^cite', 'equilibrator.gibbs.info_pages.CitePage'),
-    (r'^classic_reactions', 'equilibrator.gibbs.info_pages.ClassicReactions'),
-    (r'^compound_image', 'equilibrator.gibbs.compound_image.CompoundImage'),
-    (r'^compound', 'equilibrator.gibbs.compound_page.CompoundPage'),
-    (r'^download', 'equilibrator.gibbs.info_pages.DownloadPage'),
-    (r'^enzyme', 'equilibrator.gibbs.enzyme_page.EnzymePage'),
-    (r'^reaction', 'equilibrator.gibbs.reaction_page.ReactionPage'),
-    (r'^graph_reaction', 'equilibrator.gibbs.reaction_graph.ReactionGraph'),
-    (r'^data_refs', 'equilibrator.gibbs.data_refs_page.RefsPage'),
-    (r'^search', 'equilibrator.gibbs.search_results_page.ResultsPage'),
-    (r'^suggest', 'equilibrator.gibbs.suggest.SuggestJson'),
-    (r'^pathway$', 'equilibrator.gibbs.define_pathway_page.DefinePathwayPage'),
-    (r'^pathway/build_model', 'equilibrator.gibbs.pathway_result_page.BuildPathwayModel'),
-    (r'^pathway/results', 'equilibrator.gibbs.pathway_result_page.PathwayResultPage'),
-    (r'^robots\.txt', 'equilibrator.gibbs.info_pages.Robots'),
+urlpatterns = [
+    url(r'^$', views.MainPage),
+    url(r'^about', views.AboutPage),
+    url(r'^faq', views.FAQPage),
+    url(r'^whats_new', views.WhatsNewPage),
+    url(r'^cite', views.CitePage),
+    url(r'^classic_reactions', views.ClassicReactions),
+    url(r'^compound_image', views.CompoundImage),
+    url(r'^compound', views.CompoundPage),
+    url(r'^download', views.DownloadPage),
+    url(r'^enzyme', views.EnzymePage),
+    url(r'^reaction', views.ReactionPage),
+    url(r'^graph_reaction', views.ReactionGraph),
+    url(r'^data_refs', views.RefsPage),
+    url(r'^search', views.ResultsPage),
+    url(r'^suggest', views.SuggestJson),
+    url(r'^pathway$', views.DefinePathwayPage),
+    url(r'^pathway/build_model', views.BuildPathwayModel),
+    url(r'^pathway/results', views.PathwayResultPage),
+    url(r'^robots\.txt', views.Robots),
     # Example:
     # (r'^equilibrator/', include('equilibrator.foo.urls')),
 
-    # Uncomment the admin/doc line below and add 'django.contrib.admindocs' 
+    # Uncomment the admin/doc line below and add 'django.contrib.admindocs'
     # to INSTALLED_APPS to enable admin documentation:
     # (r'^admin/doc/', include('django.contrib.admindocs.urls')),
 
     # Uncomment the next line to enable the admin:
-    (r'^admin/', include(admin.site.urls)),
-)
+    url(r'^admin/', include(admin.site.urls)),
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
